+++ CHANGELOG.md
@@ -4,14 +4,29 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ## [Unreleased]
-- SQS MessageQueue plugin
 - remove backward compatible code after plugin feature
 - enchance the interface of Plugin
 - do lazy loading on app.loadConfig
 - app.config will use lazy loading pattern instead of checking in every function (mentioned by Jas)
-- code refactoring for executePluginPhase and asyncExecutePluginPhase (mentioned by Philip)
 - More powerful hooking for app phasing (mentioned By Philip)
 - plugin appending properties to app is not appropriate (Ted)
+
+## [ToBeReleased]
+- code refactoring for executePluginPhase and asyncExecutePluginPhase (mentioned by Philip)
+
+## [2.2.0] - 2019-07-09
+### Added
+- SQS MessageQueue plugin
+- Mongo Atlas plugin
+- AcknowledgementCenter Plugin
+- NotificationCenter Plugin
+- bind center Plugin (hooking AcknowledgementCenter and NotificationCenter)
+### Changed
+- prepare become async function
+- QueueTask plugin refactor
+  - provide `register` for hooking event and handler
+  - QueueTaskPlugin read config and regisster event and handler
+  - TODO: move QueueTask from /service to plugin
 
 ## [2.1.2] - 2019-02-26
 ### Fixed